IMPALA-8150: Fix buggy AuditingTest::TestAccessEventsOnAuthFailure
I was able to reproduce this issue by running AuditingTest individually.
Running all tests did not seem to reproduce this issue, which may be due
to the test depends on the state from other tests. Looking at the code,
the test was poorly written mainly because of two reasons:
- The Sentry config was set to an empty string, which is not a valid
  config file.
- Calling AuthorizationConfig.validateConfig() was missing.

This patch ensures validateConfig() is always called when
AuthorizationConfig instance is created.

Testing:
- Ran AuditingTest individually
- Ran all FE tests
- Ran all E2E authorization tests

Change-Id: I712697e6d5a3e171b259a781bd07de9871a29b95
Reviewed-on: http://gerrit.cloudera.org:8080/12334
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>