IMPALA-8103: In Analyzed Query use /* and */ to delimit hints.
IMPALA-5821 added the analyzed query text to the output of EXPLAIN when
explain_level is 2 or greater. Any query hits were displayed in the form

SELECT
-- +straight_join
* FROM table_name [...]

which meant that care had to be taken to embed and preserve newlines in
the EXPLAIN output. This change makes the output to be of the form

SELECT /* +straight_join */ * FROM  table_name  [...]

The  /* +straight_join */ form was chosen over  /*+straight_join */ as
it seems more readable, and is more commonly used in exiting tests.

To do this the following changes were made:
ToSqlUtils.getPlanHintsSql() was extended to take a ToSqlOptions
parameter. If this parameter indicates that the analyzed query is being
shown, then the hints are surrounded with '/*' and '*/'.
PrintUtils.wrapString() was simplified as it no longer has to preserve
newlines in a an analyzed query.

TESTING
 All end-to-end tests were run.
 The output in a few .test files was updated.
 A couple of new cases were added to ToSqlTest.java

Change-Id: I7215a4c17508e3408680a1d2bb6c3af355c78c8d
Reviewed-on: http://gerrit.cloudera.org:8080/12360
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>