IMPALA-7265: Enable caching of remote file handles by default
This changes the default value of cache_remote_file_handles
from false to true. Testing shows that this setting has a
major impact on performance for clusters that do remote HDFS
reads. Hand testing of the cache did not reveal any problems
with the semantics of caching remote file handles.

Change-Id: I2fc4a69c6bf721017f4adcdc302db9eace5135a4
Reviewed-on: http://gerrit.cloudera.org:8080/12387
Reviewed-by: Philip Zeyliger <philip@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>