IMPALA-8194: wait longer to detect JVM pause in TestPauseMonitor.
The test 'test_jvm_pause_monitor_logs_entries' stops and starts an
impalad, and confirms that that the JVM pause monitor detects the pause
by looking for a specific message in the log. In a test run the test
failed to find the correct message after sleeping for 1.2 seconds.
Because the test notes the last message that it sees in the log, we can
observe that the test would have found the correct message if it had
waited for just a few more milliseconds.

This change increases the time that the test waits to 2 seconds.

TESTING:
 Ran end-to-end tests cleanly and checked that
 test_jvm_pause_monitor_logs_entries ran OK.

Change-Id: I735c0c0ecfd3a9099c9cef332c5e79854bec7b8d
Reviewed-on: http://gerrit.cloudera.org:8080/12475
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49248
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Andrew Sherman <asherman@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
CDH-Build: Laszlo Gaal <laszlo.gaal@cloudera.com>
Quasar-L0: Laszlo Gaal <laszlo.gaal@cloudera.com>