IMPALA-8199: Fix stress test: 'No module named RuntimeProfile.ttypes'
A recent commit (IMPALA-6964) broke the stress test because it added
an import of a generated thrift value to a python file that is
included by the stress test. The stress test is intended to be able to
be run without doing a full build of Impala, but in this case the
generated thrift isn't available, leading to an import error.

The solution is to only import the thrift value in the function where
it is used, which is not called by the stress test.

Testing:
- Ran the stress test manually without doing a full build and
  confirmed that it works now.

Change-Id: I7a3bd26d743ef6603fabf92f904feb4677001da5
Reviewed-on: http://gerrit.cloudera.org:8080/12472
Reviewed-by: Thomas Marshall <thomasmarshall@cmu.edu>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49250
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
CDH-Build: Laszlo Gaal <laszlo.gaal@cloudera.com>
Quasar-L0: Laszlo Gaal <laszlo.gaal@cloudera.com>