IMPALA-5397: Set query's end_time_us_ when the operation completes
Previously, a query's end time was set only when UnregisterQuery()
was called. This was misleading because the query could have
completed earlier. After this change, if the query has a
coordinator, the end time is set when the query releases its
admission control resources. For queries without coordinators,
(e.g. DDL queries), the end time would still be set when
UnregisterQuery() is called.

Testing:
Added a test to ensure that the end time is set before the query
is closed for a query with a coordinator. It also ensures that for
a query without a coordinator, the end time is set only when the
query is closed.
Tested queries with various settings of --idle-query-timeout.
Ran exhaustive end-to-end tests and the stress test without
any failures.

Change-Id: Iaa9a1b443df3dbd95f9f297c2f923ad795b14745
(cherry picked from commit d3430ff5488caf9f547e6b9782fd029ad65e3242)
Reviewed-on: https://gerrit.sjc.cloudera.com/42220
Reviewed-by: Pooja Nilangekar <pooja.nilangekar@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>