IMPALA-8064: Increase RUNTIME_FILTER_WAIT_TIME_MS for slow builds
Several ASAN runs of test_decimal_min_max_filters failed due to
the filters not arriving in time. This was because the codegen
time for the fragments was high and hence the scan node took long
to complete. This change increases the wait time from 100000ms to
200000ms.

Testing:
Executed the test 1000 times on an ASAN build without any failures.

Change-Id: Ib18075c2a480aad1331754cfd89a383dd58b0f2e
Reviewed-on: http://gerrit.cloudera.org:8080/12560
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49116
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Quasar-L0: quasar-precommit User