IMPALA-8279: Revert IMPALA-6658 to avoid ETL performance regression.
IMPALA-6658 changed RleEncoder to have the ability to use run lengths
other than 8. It seemed that a slightly more complex RleEncoder could
save a small amount of disk space by using the longer run lengths, in
particular for bit width of 1. We now see a performance regression on a
simple ETL query.  Overall it seems that the costs of IMPALA-6658 exceed
the benefits. This change removes IMPALA-6658.

The strategy for this was that the change to rle-encoding.h, which
contains the code, was undone using 'git revert'. I removed the test
changes in rle-test.cc that rely on different encoding lengths. This
allows us to keep some useful new tests that were written as part of
IMPALA-6658

TESTING:

Ran all end-to-end tests.

Change-Id: If6bcbaf564fbbe6dc83ba3afc100b4e5ccc7af40
Reviewed-on: https://gerrit.sjc.cloudera.com/42468
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Andrew Sherman <asherman@cloudera.com>