IMPALA-8283. Order of Kudu PRIMARY KEYs can be silently ignored
When creating a Kudu table, we would use the 'PRIMARY KEY (...)' clause
to determine which columns made up the primary key, but the order of
those columns would be ignored. Thus a statement like:

CREATE TABLE (x int, y int, PRIMARY KEY (y, x)) STORED AS KUDU;

would silently create a table with an (x,y) primary key instead of a
(y,x) key. This can have substantial performance implications.

This fixes the frontend to correctly throw an error in this case.

This might be incompatible if someone was previously relying on the bug,
but I think it's worth fixing because it was clearly doing the wrong
thing.

Change-Id: I0499cee7c532db19cddac3906198d965b27ea604
Reviewed-on: http://gerrit.cloudera.org:8080/12694
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/48128
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Grant Henke <ghenke@cloudera.com>
CDH-Build: Hao Hao <hao.hao@cloudera.com>
Quasar-L0: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>