IMPALA-7982: Add host network usage to profile
This change adds host network usage to profiles. For each host that
participates in the query execution it adds the incoming and outgoing
bandwidth across all interfaces excluding the local loopback interface.
This includes all data transmitted by the host as part of the execution
of a query, other queries, and other processes running on the same
system.

The change adds tests for the added functionality to the backend and end
to end tests.

Change-Id: I2cc74f87374080a74a13b7fb6e4da44a11d828ef
Reviewed-on: http://gerrit.cloudera.org:8080/12747
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49286
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
CDH-Build: Laszlo Gaal <laszlo.gaal@cloudera.com>
Quasar-L0: Laszlo Gaal <laszlo.gaal@cloudera.com>