IMPALA-8299: Fix crash in GroupingAggregator on uninited hash table
If HashTable::Init() fails, in some cases we may try to call
HashTable::Close() on the table in GroupingAggregator::Close(), which
can cause a crash.

The solution is to set the hash table to a nullptr if Init() fails.

I also verified that the other use of HashTable, PhjBuilder, handles
this case correctly.

Testing:
- Manually verified that the crash no longer occurs if HashTable::Init
  fails.

Change-Id: Idee98d94451c9bcd958f89f6b7803ba87a5bcb8a
Reviewed-on: https://gerrit.sjc.cloudera.com/42969
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>