Remove setting of 'kudu.master_addresses' from Kudu data loading
During dataload, for tables that don't have a CREATE_KUDU section in
their table definition (eg. in functional_schema_template.sql), we
automatically generate a CREATE TABLE statement that is appropriate
for a Kudu table (eg. we specify a primary key) to create the Kudu
version of the table.

Previously, we would also set the tblproperty 'kudu.master_addresses'.
This was required at the time, but we have since added the flag
--kudu_master_hosts, which provides a default value for this
tblproperty.

The expectation during data load is that there is only a single Kudu
service, the default that is passed in to --kudu_master_hosts, so it
is no longer necessary to specify 'kudu.master_addresses' during data
load.

Testing:
- Ran a full data load and verified that the Kudu tables were loaded
  as expected.

Change-Id: If1efe49c40c98f64fa4a778044c3bb2a3d6a9cef
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/43325
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>