IMPALA-8360: Fix race conditions in thread-pool-test
There are race conditions in thread-pool-test between the caller
thread and the worker thread. Specifically, in some cases, the
worker thread seems to be slow in freeing resources. This can
lead to asserts failing because the work item has not been
freed or because the submit of a task failed when it should
not have.

This fixes the race conditions:
 - It breaks up the existing SynchronousThreadPoolTest into two
   smaller tests so that the two can't interfere with each other.
 - It adds the ability to sleep and recheck that the work item
   has been freed.

Testing:
 - Ran thread-pool-test in a loop for 20k iterations

Change-Id: Id2d5f8b677e475d8e9d6b4512e990b20bfbefaf1
Reviewed-on: http://gerrit.cloudera.org:8080/12916
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49570
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
CDH-Build: Laszlo Gaal <laszlo.gaal@cloudera.com>
Quasar-L0: Laszlo Gaal <laszlo.gaal@cloudera.com>