IMPALA-8612: Fix sporadic NPE when dropping an authorized table
In the analyze() function of DropTableOrViewStmt it's possible that
serverName_ is not set when analyzer.getTable() throws. As a result
when the Catalog executes the drop table DDL it runs into a failing
Precondition check and throws a NullPointerException when updating
user privileges. Note, to run into the NPE it's required to have
authorization enabled.

Change-Id: I70bd7ca4796b24920ee156436bf8bbc682e7d952
Reviewed-on: http://gerrit.cloudera.org:8080/13508
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62202
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Zsolt Fekete <zfekete@cloudera.com>
CDH-Build: Zsolt Fekete <zfekete@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62849
Reviewed-by: Zsolt Fekete <zfekete@cloudera.com>
Tested-by: Zsolt Fekete <zfekete@cloudera.com>