IMPALA-8561: Eliminate mtime=-1 for HDFS scan ranges (part 1)
The file handle cache uses the mtime to distinguish
different versions of a file separate. For example,
if a file at mtime=1 is overwritten with a version
at mtime=2, the old file handle from mtime=1 will
not be used for the mtime=2 version.

In some codepaths, for legacy reasons, the mtime
would be unconditionally set to -1, and this
eliminates the ability to distinguish between different
versions of files. There is no need to set the mtime to
-1. It seems to be a legacy bit of cruft.

This removes the mtime=-1 behavior for HDFS scan ranges.
It removes mtime from BufferOpts and plumbs the mtime
through the scan range codepaths separately. Local
non-HDFS files do not use the mtime, so those continue
to use mtime=-1.

Testing:
 - Passed core tests

Backport conflicts:
 - The page skipping is not present, so some overloads of
   HdfsScanNodeBase::AllocateScanRange() don't exist.

Reviewed-on: http://gerrit.cloudera.org:8080/13522
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49654
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: quasar-precommit User
Change-Id: I6b7e9bd31c43cfe7e07f53744d01a99fd04cd77b
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49697
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>