IMPALA-8561: Eliminate mtime=-1 for HDFS scan ranges (part 1)
The file handle cache uses the mtime to distinguish
different versions of a file separate. For example,
if a file at mtime=1 is overwritten with a version
at mtime=2, the old file handle from mtime=1 will
not be used for the mtime=2 version.

In some codepaths, for legacy reasons, the mtime
would be unconditionally set to -1, and this
eliminates the ability to distinguish between different
versions of files. There is no need to set the mtime to
-1. It seems to be a legacy bit of cruft.

This removes the mtime=-1 behavior for HDFS scan ranges.
It removes mtime from BufferOpts and plumbs the mtime
through the scan range codepaths separately. Local
non-HDFS files do not use the mtime, so those continue
to use mtime=-1.

Testing:
 - Passed core tests

Backport notes:
Many conflicts around changes between cdh5 and cdh6.x:
* Erasure coding boolean was added to many argument lists
* Data cache was added
* DiskIoMgr switched to reservations, changing a lot of the
  code allocating and starting scan ranges.
* Parquet scanners were moved to exec/parquet and refactored

Change-Id: I48b7ed60d6ab9104b993237b4fe23de5dc058672
Reviewed-on: http://gerrit.cloudera.org:8080/13522
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>